requirements.txt
absl-py==0.3.0
alabaster==0.7.10
anaconda-client==1.6.5
anaconda-navigator==1.6.9
anaconda-project==0.8.0
argh==0.26.2
asn1crypto==0.22.0
astroid==1.5.3
astropy==2.0.2
babel==2.5.0
backports.shutil-get-terminal-size==1.0.0
beautifulsoup4==4.6.0
bitarray==0.8.1
bkcharts==0.2
blaze==0.11.3
bleach==1.5.0
bokeh==0.12.10
boto==2.48.0
Bottleneck==1.2.1
CacheControl==0.12.3
cefpython3==57.1
certifi==2019.3.9
cffi==1.10.0
chardet==3.0.4
click==6.7
cloudpickle==0.4.0
clyent==1.2.2
colorama==0.3.9
comtypes==1.1.2
conda==4.4.8
conda-build==3.0.27
conda-verify==2.0.0
configparser==3.7.4
contextlib2==0.5.5
cryptography==2.0.3
cycler==0.10.0
Cython==0.26.1
cytoolz==0.8.2
dask==0.15.3
datashape==0.5.4
decorator==4.1.2
dill==0.2.8.2
distlib==0.2.5
distributed==1.19.1
dj-database-url==0.5.0
Django==2.1
django-filter==2.1.0
django-heroku==0.3.1
django-pandas==0.6.1
django-restframework==0.0.1
docutils==0.14
dronekit==2.9.1
entrypoints==0.2.3
enum34==1.1.6
et-xmlfile==1.0.1
fastcache==1.0.2
filelock==2.0.12
fire==0.1.3
flake8==3.5.0
Flask==1.0.2
Flask-Cors==3.0.3
Flask-SocketIO==3.1.1
future==0.17.1
futures==3.1.1
gevent==1.2.2
gitdb2==2.0.3
GitPython==2.1.9
glob2==0.5
google-images-download==2.5.0
goprocam==3.0.1
greenlet==0.4.12
gunicorn==18.0
h5py==2.7.0
heapdict==1.0.0
heroku==0.1.4
html5lib==0.9999999
idna==2.6
imageio==2.2.0
imagesize==0.7.1
imgaug==0.2.8
imutils==0.5.1
ipaddress==1.0.22
ipykernel==4.6.1
ipython==6.1.0
ipython-genutils==0.2.0
ipywidgets==7.0.0
isort==4.2.15
itsdangerous==0.24
jdcal==1.3
jedi==0.11.1
Jinja2==2.10
jsonpickle==1.2
jsonschema==2.6.0
jupyter-client==5.1.0
jupyter-console==5.2.0
jupyter-core==4.3.0
jupyterlab==0.27.0
jupyterlab-launcher==0.4.0
Keras==2.2.4
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
keyboard==0.13.2
Kivy==1.10.0
Kivy-examples==1.10.0
Kivy-Garden==0.1.4
kivy.deps.angle==0.1.6
kivy.deps.glew==0.1.9
kivy.deps.glew-dev==0.1.9
kivy.deps.gstreamer==0.1.12
kivy.deps.gstreamer-dev==0.1.12
kivy.deps.sdl2==0.1.17
kivy.deps.sdl2-dev==0.1.17
kiwisolver==1.0.1
lazy-object-proxy==1.3.1
llvmlite==0.20.0
locket==0.2.0
lockfile==0.12.2
lxml==4.1.0
mail-parser==3.9.3
Markdown==2.6.11
MarkupSafe==1.0
matplotlib==3.0.2
mccabe==0.6.1
menuinst==1.4.10
mistune==0.7.4
monotonic==1.5
mpmath==0.19
msgpack-python==0.4.8
multipledispatch==0.4.9
navigator-updater==0.1.0
nbconvert==5.3.1
nbformat==4.4.0
networkx==2.0
nltk==3.2.4
nose==1.3.7
notebook==5.0.0
numba==0.35.0+10.g143f70e
numexpr==2.6.2
numpy==1.16.2
numpydoc==0.7.0
odo==0.5.1
olefile==0.44
opencv-python==3.4.0
openpyxl==2.4.8
packaging==16.8
pandas==0.20.3
pandocfilters==1.4.2
parso==0.1.1
partd==0.3.8
path.py==10.3.1
pathlib2==2.3.0
pathtools==0.1.2
patsy==0.4.1
pep8==1.7.0
pickleshare==0.7.4
Pillow==6.0.0
pkginfo==1.4.1
plotly==2.6.0
ply==3.10
progress==1.3
prompt-toolkit==1.0.15
protobuf==3.5.1
psutil==5.5.0
psycopg2==2.8.3
py==1.4.34
pyarrow==0.9.0
pycodestyle==2.3.1
pycosat==0.6.3
pycparser==2.18
pycrypto==2.6.1
pycurl==7.43.0
pyflakes==1.6.0
pygame==1.9.3
Pygments==2.2.0
pylint==1.7.4
pymavlink==2.3.4
pyodbc==4.0.17
pyOpenSSL==17.2.0
pyparsing==2.2.0
pypiwin32==223
PyQt5==5.12.1
PyQt5-sip==4.19.15
pyreadline==2.1
pyserial==3.4
PySide2==5.11.2
PySocks==1.6.7
pytest==3.2.1
python-dateutil==2.8.0
python-dotenv==0.9.1
python-engineio==3.0.0
python-socketio==2.1.2
pytz==2017.2
PyWavelets==0.5.2
pywin32==223
PyYAML==3.12
pyzmail36==1.0.4
pyzmq==16.0.2
qgmap==0.2
QtAwesome==0.4.4
qtconsole==4.3.1
QtPy==1.3.1
requests==2.18.4
rope==0.10.5
ruamel-yaml==0.11.14
scikit-image==0.15.0
scikit-learn==0.19.1
scipy==0.19.1
seaborn==0.8
selenium==3.141.0
Shapely==1.6.4.post1
shiboken2==5.12.0a1.dev1540884599
simplegeneric==0.8.1
simplejson==3.16.0
singledispatch==3.4.0.3
sip==4.19.8
six==1.12.0
slidingwindow==0.0.13
smmap2==2.0.3
snowballstemmer==1.2.1
sortedcollections==0.5.3
sortedcontainers==1.5.7
Sphinx==1.6.3
sphinxcontrib-websupport==1.0.1
spyder==3.2.4
SQLAlchemy==1.1.13
statsmodels==0.8.0
sympy==1.1.1
tables==3.4.2
tabulate==0.8.2
tblib==1.3.2
tensorflow==1.5.0
tensorflow-gpu==1.5.0
tensorflow-tensorboard==1.5.1
tensorpack==0.8.6
termcolor==1.1.0
testpath==0.3.1
toolz==0.8.2
tornado==4.5.2
tqdm==4.24.0
traitlets==4.3.2
typing==3.6.2
unicodecsv==0.14.1
urllib3==1.22
watchdog==0.8.3
wcwidth==0.1.7
webcam-streamer==1.0.5
webencodings==0.5.1
Werkzeug==0.14.1
whitenoise==4.1.2
widgetsnbextension==3.0.2
win-inet-pton==1.0.1
win-unicode-console==0.5
wincertstore==0.2
wrapt==1.10.11
xlrd==1.1.0
XlsxWriter==1.0.2
xlwings==0.11.4
xlwt==1.3.0
zict==0.1.3
